Extract dropdowns into their own components.
